`--keep` and `cache` (#4)
This adds a --keep command-line options to the scripts. Binaries will be kept for this many seconds before a new download is attempted. The default for the option is 0, meaning that downloads will always be attempted, which is backwards compatible with prior versions. The --keep option is relayed by a cache input, which defaults to 1d (one day).

Closes #2